[ui] Split server-logger and client-logger
