+++ app/src/app/api/legal-units/[id]/primary/route.ts
@@ -1,15 +1,15 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@/lib/supabase/server";
-import { createServerLogger } from "@/lib/logger";
+import { createServerLogger } from "@/lib/server-logger";
 
 export async function POST(
   request: Request,
   { params: { id } }: { readonly params: { id: string } }
 ) {
+  const logger = await createServerLogger();
   try {
     const enterprise: { unit_id: number } = await request.json();
     const legalUnitId = parseInt(id, 10);
-    const logger = await createServerLogger();
 
     if (!enterprise?.unit_id || !legalUnitId) {
       logger.error(
