+++ app/src/app/establishments/[id]/page.tsx
@@ -5,7 +5,8 @@ import { Button } from "@/components/ui/button";
 import { createClient } from "@/lib/supabase/server";
 import { revalidatePath } from "next/cache";
 import { InfoBox } from "@/components/info-box";
-import { createServerLogger } from "@/lib/logger";
+
+import { createServerLogger } from "@/lib/server-logger";
 
 export default async function EstablishmentGeneralInfoPage({
   params: { id },
