+++ app/src/app/global-error-reporter.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useEffect } from "react";
-import logger from "@/lib/logger";
+import logger from "@/lib/client-logger";
 
 export default function GlobalErrorReporter() {
   useEffect(() => {
